Consumer.Tests/RabbitMqConsumerTests.cs
using Castle.Core.Configuration;
using Moq;
using RabbitMQ.Client;
using Xunit;

namespace Consumer.Tests
{
    // In the same directory as your RabbitMqConsumer.cs file
    public class RabbitMqConsumerTests
    {
        private readonly Mock<IRabbitMqConnectionService> _rabbitMqConsumerServiceMock;
        private readonly Mock<IConfiguration> _configurationMock;
        private readonly Mock<RabbitMqConsumer> _rabbitMqConsumerMock;
        private Mock<IModel> _channel;

        public RabbitMqConsumerTests()
        {
            _rabbitMqConsumerServiceMock = new Mock<IRabbitMqConnectionService>();  
            _configurationMock = new Mock<IConfiguration>();
            _rabbitMqConsumerMock = new  Mock<RabbitMqConsumer>();
            _channel = new Mock<IModel>();

            _rabbitMqConsumerServiceMock.Setup(s => s.CreateChannel()).Returns(_channel.Object);

        }

        [Fact]
        public async Task TestStartConsumingAsync()
        {
            // Arrange
            var mockConnectionService = new Mock<IRabbitMqConnectionService>();
            var mockChannel = new Mock<IModel>();

            // Setup the mock connection service to return the mock channel


            // Act



            // Act
            // await consumer.StartConsumingAsync(CancellationToken.None);

            // // Assert
            // // Check if the channel is created
            // Assert.NotNull(consumer._channel);

            // // Verify if the queue was declared
            // Assert.NotNull(consumer._consumerTag);
        }

        //[Fact]
        //public async Task TestPauseQueue()
        //{
        //    // Arrange
        //    var mockConnectionService = new Mock<IRabbitMqConnectionService>();
        //    var mockConfig = new Mock<IConfiguration>();
        //    var consumer = new RabbitMqConsumer(mockConnectionService.Object);

        //    // Act
        //    await consumer.PauseQueue();

        //    // Assert
        //    // TODO: Implement assertions based on the actual behavior of PauseQueue
        //}

        //[Fact]
        //public async Task TestResumeQueue()
        //{
        //    // Arrange
        //    var mockConnectionService = new Mock<IRabbitMqConnectionService>();
        //    var mockConfig = new Mock<IConfiguration>();
        //    var consumer = new RabbitMqConsumer(mockConnectionService.Object);

        //    // Act
        //    consumer.ResumeQueue();

        //    // Assert
        //    // TODO: Implement assertions based on the actual behavior of ResumeQueue
        //}

        //[Fact]
        //public async Task TestSetSchedule()
        //{
        //    // Arrange
        //    var mockConnectionService = new Mock<IRabbitMqConnectionService>();
        //    var mockConfig = new Mock<IConfiguration>();
        //    var consumer = new RabbitMqConsumer(mockConnectionService.Object);

        //    // Act
        //    consumer.SetSchedule(new DateTime(2023, 1, 1), new DateTime(2023, 1, 2));

        //    // Assert
        //    // TODO: Implement assertions based on the actual behavior of SetSchedule
        //}
    }
}